Add basic functionality to etch a sketch
This etch a sketch website has basic functionality including a slider for adjusting grid size that resets the etch a sketch automatically every time the square sizes are changed. The slider only contains multiples that have 0 remainder when divided into 960, the amount of pixels of with and height that the grid contains.